requirements.txt
wandb
hydra-core
jaxtyping
hickle
numpy
scipy
numpy
matplotlib
tqdm
opencv-python
pandas
pillow
einops
hydra_colorlog
torch-tb-profiler
tensorboard
autoroot
datasets
typer
hdf5plugin
timm
autopep8
imageio
imageio-ffmpeg
ipdb
debugpy
torchinfo
accelerate
submitit
diffusers
segment-anything-hq
git+https://github.com/alexanderswerdlow/image_utils.git